Added proposal_name secondary index
This addition makes it "much" easier to retrieve the all the votes for a
given proposal.

Instead of having to know the exact key information, one can now use a
simpler name value. However, the upper bound value must still be name + 1
since the upper value is exclusive and makes it hard to query just for
a given one.